requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase>=2.3.4
pydantic==2.6.1
python-jose[cryptography]==3.3.0
httpx>=0.27.0
python-multipart==0.0.9
